Comparison between signed and unsigned
Review URL: http://codereview.chromium.org/6411
Patch from icefox.


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@3046 0039d316-1c4b-4281-b951-d872f2087c98

+++ webkit/glue/npruntime_util.cc
@@ -52,7 +52,7 @@ bool DeserializeNPIdentifier(const Pickle& pickle, void** pickle_iter,
     int data_len;
     if (!pickle.ReadData(pickle_iter, &data, &data_len))
       return false;
-    DCHECK_EQ(data_len, strlen(data) + 1);
+    DCHECK_EQ((static_cast<size_t>(data_len)), strlen(data) + 1);
     *identifier = NPN_GetStringIdentifier(data);
   } else {
     int number;
